Branch code tidy.
