--- tests/cfgspec/test_header
@@ -1,1 +1,0 @@
-../lib/bash/test_header
